requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.45.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.0
pypdf==3.16.2
faiss-cpu==1.12.0
numpy==2.3.2